pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "team_memory"
version = "0.1.0"
description = "Team experience database service based on MCP protocol"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "pgvector>=0.3.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "alembic>=1.13.0",
    "httpx>=0.27.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
redis = [
    "redis[hiredis]>=5.0.0",
]
local-embedding = [
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "testcontainers[postgres]>=4.0.0",
]

[project.scripts]
team-memory = "team_memory.server:main"
team-memory-web = "team_memory.web.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/team_memory"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
